Api calls without viewmodel

+++ app/src/main/java/com/example/myapplication/ImageLoad.kt
@@ -1,5 +1,5 @@
 package com.example.myapplication
 
-data class ImageLoad(val imageUrl: String,var rating: Int)
+data class ImageLoad(val imageUrl: String,var rating: Float)
 {}
